Lisätty GUI

+++ minigolf/src/main/java/fi/yussiv/minigolf/domain/Position.java
@@ -6,19 +6,19 @@ package fi.yussiv.minigolf.domain;
  * @author jkvoipio
  */
 public class Position {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Position(int x, int y) {
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
